Change the scanning to occur during the handling of events on eventbus
Change how the scan is invoked in order to occur within a handler of a
specific scanning event.  The scan event is posted onto the event bus.
The policy is also executed as an event on the event bus.  The scanning
event handler posts the policy check event to be handled by the
appropriate handler.